Revert "separating builder/marketplace"
This reverts commit 9ef0efb1a8d4c3a7b32808e01d48d2da1b17e0b1.
