+++ apps/builder/app/builder/features/sidebar-left/panels/store/item-panel.tsx
@@ -1,13 +1,13 @@
 import { Button, Flex, theme } from "@webstudio-is/design-system";
 import { Iframe } from "./iframe";
-import type { MarketplaceItem } from "./types";
+import type { StoreItem } from "./types";
 import { ChevronLeftIcon } from "@webstudio-is/icons";
 
 export const ItemPanel = ({
   item,
   onOpenChange,
 }: {
-  item: MarketplaceItem;
+  item: StoreItem;
   onOpenChange: (isOpen: boolean) => void;
 }) => {
   return (
