+++ apps/builder/app/builder/features/sidebar-left/panels/store/panel.tsx
@@ -2,7 +2,7 @@ import { PluginIcon } from "@webstudio-is/icons";
 import { Flex } from "@webstudio-is/design-system";
 import type { TabContentProps } from "../../types";
 import { Header, CloseButton } from "../../header";
-import { Marketplace } from "./marketplace";
+import { Store } from "./store";
 import { ItemDialog } from "./item-dialog";
 import { ItemPanel } from "./item-panel";
 import { useActiveItem } from "./utils";
@@ -29,7 +29,7 @@ export const TabContent = ({ onSetActiveTab }: TabContentProps) => {
       {component === "panel" && activeItem ? (
         <ItemPanel item={activeItem} onOpenChange={handleOpenChange} />
       ) : (
-        <Marketplace />
+        <Store />
       )}
     </Flex>
   );
