--- apps/builder/app/routes/_builder-shared.tsx
@@ -1,124 +1,0 @@
-import { useRouteError, isRouteErrorResponse } from "@remix-run/react";
-import { redirect, type LoaderArgs, json } from "@remix-run/node";
-import { loadBuildByProjectId } from "@webstudio-is/project-build/index.server";
-import { db as domainDb } from "@webstudio-is/domain/index.server";
-import { db } from "@webstudio-is/project/index.server";
-import {
-  AuthorizationError,
-  authorizeProject,
-} from "@webstudio-is/trpc-interface/index.server";
-import { loadAssetsByProject } from "@webstudio-is/asset-uploader/index.server";
-import { createContext } from "~/shared/context.server";
-import { ErrorMessage } from "~/shared/error";
-import { sentryException } from "~/shared/sentry";
-import { loginPath } from "~/shared/router-utils";
-import type { BuilderProps } from "~/builder";
-
-export const loader = async ({
-  params,
-  request,
-}: LoaderArgs): Promise<BuilderProps> => {
-  const context = await createContext(request);
-
-  try {
-    if (params.projectId === undefined) {
-      throw new Error("Project id undefined");
-    }
-
-    const url = new URL(request.url);
-
-    const start = Date.now();
-    const project = await db.project.loadById(params.projectId, context);
-
-    const authPermit =
-      (await authorizeProject.getProjectPermit(
-        {
-          projectId: project.id,
-          // At this point we already knew that if project loaded we have at least "view" permit
-          // having that getProjectPermit is heavy operation we can skip check "view" permit
-          permits: ["own", "admin", "build"] as const,
-        },
-        context
-      )) ?? "view";
-
-    if (project === null) {
-      throw new Error(`Project "${params.projectId}" not found`);
-    }
-
-    const devBuild = await loadBuildByProjectId(project.id);
-
-    const assets = await loadAssetsByProject(project.id, context);
-
-    const currentProjectDomainsResult = await domainDb.findMany(
-      { projectId: project.id },
-      context
-    );
-
-    const domains = currentProjectDomainsResult.success
-      ? currentProjectDomainsResult.data
-          .filter((projectDomain) => projectDomain.verified)
-          .filter((projectDomain) => projectDomain.domain.status === "ACTIVE")
-          .map((projectDomain) => projectDomain.domain.domain)
-      : [];
-
-    const end = Date.now();
-
-    const diff = end - start;
-
-    // we need to log timings to figure out how to speed up loading
-    // eslint-disable-next-line no-console
-    console.log(`Project ${project.id} is loaded in ${diff}ms`);
-
-    const authToken = url.searchParams.get("authToken") ?? undefined;
-
-    const { userPlanFeatures } = context;
-    if (userPlanFeatures === undefined) {
-      throw new Error("User plan features are not defined");
-    }
-
-    if (project.userId === null) {
-      throw new AuthorizationError("Project must have project userId defined");
-    }
-
-    return {
-      project,
-      domains,
-      build: devBuild,
-      assets: assets.map((asset) => [asset.id, asset]),
-      authToken,
-      authPermit,
-      userPlanFeatures,
-    };
-  } catch (error) {
-    if (error instanceof AuthorizationError) {
-      // try to login user if he is not logged in
-      if (context.authorization.userId === undefined) {
-        const url = new URL(request.url);
-
-        throw redirect(
-          loginPath({
-            returnTo: url.pathname,
-          })
-        );
-      }
-
-      const FORBIDDEN = 403;
-
-      throw json({ message: error.message }, { status: FORBIDDEN });
-    }
-
-    throw error;
-  }
-};
-
-export const ErrorBoundary = () => {
-  const error = useRouteError();
-  sentryException({ error });
-  const message = isRouteErrorResponse(error)
-    ? error.data.message ?? error.data
-    : error instanceof Error
-    ? error.message
-    : String(error);
-
-  return <ErrorMessage message={message} />;
-};
