--- apps/builder/app/routes/marketplace.ts
@@ -1,2 +1,0 @@
-export * from "./builder";
-export { default } from "./builder";
